build err
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@14814 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_cpp.cpp
@@ -1295,7 +1295,7 @@ SWITCH_DECLARE(switch_status_t) CoreSession::process_callback_result(char *resul
 				p++;
 				if (*p == '+' || *p == '-') {
 					int step;
-					int target;
+					int64_t target;
 					if (!(step = atoi(p))) {
 						step = 1000;
 					}
@@ -1307,7 +1307,7 @@ SWITCH_DECLARE(switch_status_t) CoreSession::process_callback_result(char *resul
 						target = 0;
 					}
 
-					switch_log_printf(SWITCH_CHANNEL_SESSION_LOG(session), SWITCH_LOG_DEBUG, "seek to position %d\n", target);
+					switch_log_printf(SWITCH_CHANNEL_SESSION_LOG(session), SWITCH_LOG_DEBUG, "seek to position %d\n", (uint32_t)target);
 					switch_core_file_seek(fhp, &pos, target, SEEK_SET);
 
 				} else {
